Create CurrencyFormatter utils
