+++ lib/bloc/user_profile/currency.dart
@@ -1,3 +1,4 @@
+import 'package:breez/utils/currency_formatter.dart';
 import 'package:fixnum/fixnum.dart';
 import 'package:intl/intl.dart';
 import 'package:intl/number_symbols.dart';
@@ -63,30 +64,7 @@ class Currency extends Object {
 }
 
 class _CurrencyFormatter {
-  static final formatter = _defineFormatter();
-
-  static NumberFormat _defineFormatter() {
-    numberFormatSymbols['space-between'] = NumberSymbols(
-      NAME: "zz",
-      DECIMAL_SEP: '.',
-      GROUP_SEP: '\u00A0',
-      PERCENT: '%',
-      ZERO_DIGIT: '0',
-      PLUS_SIGN: '+',
-      MINUS_SIGN: '-',
-      EXP_SYMBOL: 'e',
-      PERMILL: '\u2030',
-      INFINITY: '\u221E',
-      NAN: 'NaN',
-      DECIMAL_PATTERN: '#,##0.###',
-      SCIENTIFIC_PATTERN: '#E0',
-      PERCENT_PATTERN: '#,##0%',
-      CURRENCY_PATTERN: '\u00A4#,##0.00',
-      DEF_CURRENCY_CODE: 'AUD',
-    );
-    final formatter = NumberFormat('###,###.##', 'space-between');
-    return formatter;
-  }
+  static final formatter = CurrencyFormatter.formatter;
 
   String format(satoshies, Currency currency,
       {bool addCurrencySuffix = true,
