+++ lib/utils/currency_formatter.dart
@@ -1,0 +1,30 @@
+import 'package:intl/intl.dart';
+import 'package:intl/number_symbols.dart';
+import 'package:intl/number_symbols_data.dart';
+
+class CurrencyFormatter {
+  static final NumberFormat formatter = _defineFormatter();
+
+  static NumberFormat _defineFormatter() {
+    numberFormatSymbols['space-between'] = NumberSymbols(
+      NAME: "zz",
+      DECIMAL_SEP: '.',
+      GROUP_SEP: '\u00A0',
+      PERCENT: '%',
+      ZERO_DIGIT: '0',
+      PLUS_SIGN: '+',
+      MINUS_SIGN: '-',
+      EXP_SYMBOL: 'e',
+      PERMILL: '\u2030',
+      INFINITY: '\u221E',
+      NAN: 'NaN',
+      DECIMAL_PATTERN: '#,##0.###',
+      SCIENTIFIC_PATTERN: '#E0',
+      PERCENT_PATTERN: '#,##0%',
+      CURRENCY_PATTERN: '\u00A4#,##0.00',
+      DEF_CURRENCY_CODE: 'AUD',
+    );
+    final formatter = NumberFormat('###,###.##', 'space-between');
+    return formatter;
+  }
+}
